Minor doc fixes
Links to Bevy types don't work on docs.rs. The workaround is to
use actual url to docs.rs. That seems tedious and fragile, so
we'll just remove the links and inser the word "Bevy" so that
people know where to look.

See https://github.com/rust-lang/docs.rs/issues/1588